fix(v0.9.2): resolve CI segmentation faults in array function tests
Problem:
- CI tests failed with segmentation fault (exit code 35584) on Ubuntu/Linux
- Tests: func/array_func.cb and func_type_check/array_types.cb
- Tests passed locally on macOS but failed in CI environment
- Array functions returning arrays caused crashes when array literals were passed as arguments

Root Causes:
1. Array literal arguments: assign_array_parameter created references to temporary variables
   that went out of scope, leaving invalid pointers
2. Duplicate function calls: processArrayDeclaration and execute_array_decl both called
   evaluate() for function call initializations
3. Invalid array data: Function returns had array_size=0 due to empty array_values

Solutions:
1. Modified assign_array_parameter to detect array literals and deep copy data instead
   of creating references (manager.cpp)
2. Added safety check using var.is_assigned to prevent duplicate evaluate() calls
   (array_declaration.cpp)
3. Clean up: Removed 233-line commented code block that was causing file issues
4. Added comprehensive debug logging for troubleshooting

Technical Changes:
- src/backend/interpreter/managers/variables/manager.cpp:
  * assign_array_parameter: Detect literals vs variables, copy data for literals
- src/backend/interpreter/executors/declarations/array_declaration.cpp:
  * Added var.is_assigned check before evaluate() to prevent duplicate calls
  * Removed large commented-out code block
- .github/workflows/ci.yml:
  * Removed debug-build-test job
  * Added clean checkout configuration

Testing:
- All 2409 tests pass (2379 integration + 30 unit)
- array_func.cb: Functions correctly return arrays
- array_types.cb: Array literals and variables work as function arguments
- Expected to resolve CI segmentation faults on Linux/Ubuntu

Related: v0.9.2 array reference passing and return value features

+++ src/backend/interpreter/executors/statement_executor.cpp
@@ -899,7 +899,8 @@ void StatementExecutor::execute_ternary_variable_initialization(
 
     // 三項演算子の条件を評価
     int64_t condition = interpreter_.evaluate(ternary_node->left.get());
-    printf("DEBUG: Ternary condition = %lld\n", condition);
+    printf("DEBUG: Ternary condition = %lld\n",
+           static_cast<long long>(condition));
 
     // 条件に基づいて選択される分岐を決定
     const ASTNode *selected_branch =
